requirements.txt
sqlalchemy~=1.3.18
requests~=2.24.0
psycopg2-binary
urllib3~=1.25.9
python-dateutil~=2.8.1
netaddr~=0.8.0
pandas~=1.1.0